Implement scale transforms.
Since gragrapy is now telling mpl where the axis limits are, it needs to
keep track of everything drawn. That means for geom.bar, scale.y needs
to know that the bars draw at 0. Since scale.y doesn't even know we have
geom.bar, we need to put the 0 into the data frame as ymin. In turn,
that means scales need to be able to apply to multiple columns at once.