Start refactoring plotting.
Faceting will work by adding a 'facet' column to the dataset, and things
that need to split up facets can split them up themselves.

Hoping to be able to rewrite Plot.make() mostly as a series of mapped
functions. I could do that with the all_mapped block now, except that
this has exposed a bug, in that the all_statted block references aes
from the all_mapped block.