Map position scales before stats, not after.
I want stat.bin to work with discrete scales, and this is part of a
movement towards that.

We now train scales in two stages: first we train the dataspace mapping,
and then (after the stat) we train the guides (axes and legends). This
way the guides are trained on the data that comes out of the stats, not
the data that goes into them.